改善chatpdf的功能
