+++ version
@@ -1,5 +1,5 @@
 {
-  "version": 2.67,
+  "version": 2.68,
   "show_feature": true,
-  "new_feature": "现可通过输入区更新临时api-key <-> 增强多线程稳定性（涉及代码解析、PDF翻译、自译解等） <-> 修复Token计数错误（解决PDF翻译的分割不合理的问题） <-> 如果一键更新失败，可前往github手动更新"
+  "new_feature": "改善理解pdf（chatpdf）功能 <-> 如果一键更新失败，可前往github手动更新"
 }
